refactor: headerBar
