+++ assets/css/global.css
@@ -1,4 +1,5 @@
 @import "./transition.css";
+@import "@gits-id/avatar/dist/style.css";
 
 body {
   @apply bg-gray-100;
